requirements.txt
# Steel Fabrication — Python dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.25.0
google-cloud-secret-manager>=2.16.0
google-cloud-firestore>=2.13.0
google-auth>=2.23.0
gspread>=5.12.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
